refactor: 계층 표시 파일 추가(#19)
