--- src/main/java/demago/gamjappang/user/adapter/in/web/dto/LogoutRequest.java
@@ -1,8 +1,0 @@
-package demago.gamjappang.user.adapter.in.web.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LogoutRequest(
-
-        @NotBlank String refreshToken
-) {}
