--- src/main/java/demago/gamjappang/user/adapter/in/web/dto/TokenResponse.java
@@ -1,13 +1,0 @@
-package demago.gamjappang.user.adapter.in.web.dto;
-
-import demago.gamjappang.user.application.port.in.result.TokenResult;
-
-public record TokenResponse(
-
-        String accessToken,
-        String refreshToken
-) {
-    public static TokenResponse from(TokenResult tokenResult) {
-        return new TokenResponse(tokenResult.accessToken(), tokenResult.refreshToken());
-    }
-}
