+++ src/main/java/demago/gamjappang/user/infrastructure/adapter/in/web/dto/request/JoinRequest.java
@@ -1,0 +1,17 @@
+package demago.gamjappang.user.infrastructure.adapter.in.web.dto.request;
+
+import demago.gamjappang.user.applicationcore.port.in.command.JoinCommand;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
+public record JoinRequest(
+
+        @NotBlank @Size(max = 50) String username,
+        @NotBlank @Email @Size(max = 120) String email,
+        @NotBlank @Size(min = 8, max = 72) String password
+) {
+    public JoinCommand toCommand() {
+        return new JoinCommand(username, email, password);
+    }
+}
