+++ src/main/java/demago/gamjappang/user/infrastructure/adapter/in/web/dto/request/LogoutRequest.java
@@ -1,0 +1,8 @@
+package demago.gamjappang.user.infrastructure.adapter.in.web.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LogoutRequest(
+
+        @NotBlank String refreshToken
+) {}
