+++ src/main/java/demago/gamjappang/user/infrastructure/adapter/in/web/dto/request/VerifyRequest.java
@@ -1,0 +1,16 @@
+package demago.gamjappang.user.infrastructure.adapter.in.web.dto.request;
+
+import demago.gamjappang.user.applicationcore.port.in.command.VerifyEmailCommand;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
+public record VerifyRequest(
+
+        @NotBlank @Email String email,
+        @NotBlank @Size(min = 4, max = 10) String signupCode
+) {
+    public VerifyEmailCommand toCommand() {
+        return new VerifyEmailCommand(email, signupCode);
+    }
+}
